refactor: group media setters and getters in dedicated interfaces
