+++ src/main/java/de/unistuttgart/einf/moviemanager/model/TopLevelMedia.java
@@ -8,4 +8,11 @@ package de.unistuttgart.einf.moviemanager.model;
  */
 public sealed interface TopLevelMedia extends Media permits Movie, Series {
 
+	/**
+	 * Sets the category of the media.
+	 *
+	 * @param category the category
+	 */
+	void setCategory(Category category);
+
 }
